requirements.txt
pymysql
Flask>=2.2.5
SQLAlchemy>=2.0.21
Flask-SQLAlchemy>=3.1.1
Flask-WTF>=1.1.1
WTForms>=3.0.1
Flask-Bcrypt>=1.0.1
Flask-Login>=0.6.2
gunicorn>=21.2.0
pyodbc>=4.0.39
email-validator
Flask
mysql-connector-python
Flask-WTF

